requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
spacy==3.1.1
streamlit==0.84.0
PyPDF2==1.26.0
python-docx==0.8.11
requests==2.25.1